+++ linphone-app/assets/languages/zh_CN.ts
@@ -2026,7 +2026,7 @@
     <message>
         <source>checkForUpdates</source>
         <extracomment>&apos;Check for updates&apos; : Item menu for checking updates</extracomment>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">检测软件更新</translation>
     </message>
     <message>
         <source>recordings</source>
@@ -2056,7 +2056,7 @@
     <message>
         <source>checkForUpdates</source>
         <extracomment>&apos;Check for updates&apos; : Item menu for checking updates</extracomment>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">检测软件更新</translation>
     </message>
     <message>
         <source>recordings</source>