+++ linphone-app/src/components/settings/SettingsModel.hpp
@@ -85,6 +85,7 @@ class SettingsModel : public QObject {
 	Q_PROPERTY(QString ringPath READ getRingPath WRITE setRingPath NOTIFY ringPathChanged)
 	
 	Q_PROPERTY(bool echoCancellationEnabled READ getEchoCancellationEnabled WRITE setEchoCancellationEnabled NOTIFY echoCancellationEnabledChanged)
+	Q_PROPERTY(int echoCancellationCalibration READ getEchoCancellationCalibration NOTIFY echoCancellationCalibrationChanged)
 	
 	Q_PROPERTY(bool showAudioCodecs READ getShowAudioCodecs WRITE setShowAudioCodecs NOTIFY showAudioCodecsChanged)
 	
@@ -369,6 +370,7 @@ public:
 	
 	bool getEchoCancellationEnabled () const;
 	void setEchoCancellationEnabled (bool status);
+	int getEchoCancellationCalibration()const;
 	
 	Q_INVOKABLE void startEchoCancellerCalibration();
 	
@@ -780,6 +782,7 @@ signals:
 	
 	void echoCancellationEnabledChanged (bool status);
 	void echoCancellationStatus(int status, int msDelay);
+	void echoCancellationCalibrationChanged();
 	
 	void showAudioCodecsChanged (bool status);