+++ CHANGELOG.md
@@ -24,6 +24,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - '[ui] systray_notification_blink' : option to activate/deactivate the blinking systray on unread notifications.
 - '[ui] systray_notification_global' : option to display notification number from all accounts or only selected.
 - '[ui] systray_notification_filtered' : option to filter the notification number (not count if chat room is muted).
+- Keyboard shortcuts:
+    * 'Ctrl+Shift+W' (or V): accept with video the last incoming call.
+    * 'Ctrl+Shift+A': accept without video the last incoming call.
+    * 'Ctrl+Shift+D': terminate the last call.
+    * 'Ctrl+Shift+E': Enable/disable echo cancellation.
+    * 'Ctrl+Shift+L': Unmute/Mute speaker.
+    * 'Ctrl+Shift+M': Unmute/Mute microphone.
+- Request application focus when hovering a call notification.
 
 ## 5.2.1 - 2024-02-01